pyproject.toml
[tool.poetry]
name = "notify"
version = "0.1.0"
description = "A simple notification system based with FCM and FastAPI"
authors = ["flavien-hugs <flavienhugs@pm.me>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "3.12.3"
pydantic-settings = "2.5.2"
uvloop = "0.20.0"
fastapi = {version = "0.115.0", extras = ["standard"]}
firebase-admin = "6.6.0"
beanie = "1.26.0"
fastapi-pagination = "0.12.21"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.8.0"
black = "25.1.0"
bandit = "1.8.3"
flake8 = "7.2.0"
flake8-pyproject = "1.2.3"
isort = "5.13.2"


[tool.poetry.group.test.dependencies]
pytest = "7.3.1"
pytest-dotenv = "0.5.2"
pytest-mock = "3.14.0"
pytest-cov = "4.1.0"
coverage = "7.6.12"
faker = "30.10.0"

[tool.poetry.scripts]
app = 'src.cli:app'

[tool.flake8]
max-line-length = 130
exclude = [
    ".git",
    "src/common",
    "__pycache__"
]

[tool.black]
line-length = 130
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.docker
    | \*.egg-info
    | _build
    | buck-out
    | build
    | dist
    | src/common
  )/
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
